Added "tap to hide" behavior to series legends. Tapping on an entry in the legend will hide its related data on the chart.
PiperOrigin-RevId: 194322108

+++ charts_common/lib/src/chart/cartesian/axis/axis.dart
@@ -217,7 +217,7 @@ abstract class Axis<D> extends ImmutableAxis<D> implements LayoutView {
         providedTicks.remove(tick);
       } else {
         // Animate out ticks that do not exist any more.
-        animatedTick.animateOut(_scale[animatedTick.value].toDouble());
+        animatedTick.animateOut(_scale[animatedTick.value]);
       }
     }
 
@@ -225,7 +225,7 @@ abstract class Axis<D> extends ImmutableAxis<D> implements LayoutView {
     providedTicks?.forEach((tick) {
       final animatedTick = new AxisTicks<D>(tick);
       if (_previousScale != null) {
-        animatedTick.animateInFrom(_previousScale[tick.value].toDouble());
+        animatedTick.animateInFrom(_previousScale[tick.value]);
       }
       _axisTicks.add(animatedTick);
     });
